Refactor.
* Use snake_case in file names;
* Fix some method names;
* Remove redundant parentheses from methods;
* Rename GemInfo to MyGem;
* Use constructor for initializing MyGem properties;
* Convert ArgumentsParser to module instead of class.
