--- lib/gemInfo.rb
@@ -1,7 +1,0 @@
-class GemInfo
-  @name = ''
-  @info = ''
-
-  attr_accessor :name
-  attr_accessor :info
-end
