+++ lib/my-gem.rb
@@ -1,0 +1,10 @@
+class MyGem
+  
+  def initialize(name, info)
+    @name = name
+    @info = info
+  end
+
+  attr_accessor :name
+  attr_accessor :info
+end
